feat: implement TACT manifest integration and fix ngdp-cache race condition
This commit adds comprehensive TACT manifest integration to enable FileDataID-based
file lookups in CASC storage, and resolves a critical race condition in the cache system.

### TACT Manifest Integration

**Core Implementation:**
- Add TactManifests struct for managing root and encoding manifests
- Support BLTE decompression for manifest files
- Enable FileDataID → CKey → EKey mapping with locale/platform filtering
- Add listfile support for filename → FileDataID resolution
- Implement comprehensive caching with BTreeMap for stable iteration

**CLI Integration:**
- New storage commands: extract-by-id, extract-by-name, load-manifests
- Support for both FileDataID and filename-based file extraction
- Visual progress indicators and multiple output formats (text, JSON, BPSV)
- Automatic filename sanitization for safe output paths

**Files Added:**
- casc-storage/src/manifest/ - TACT manifest integration module
- casc-storage/tests/tact_manifest_tests.rs - Comprehensive test suite
- ngdp-client/src/commands/listfile.rs - Listfile management commands

**Key Features:**
- FileDataID-based file extraction from CASC storage
- Filename-based extraction using community listfiles
- Platform and locale filtering for multi-platform game builds
- Memory-efficient caching with configurable options

### Race Condition Fix

**Problem:** test_cache_ttl_and_validity failing intermittently with CacheEntryNotFound

**Root Cause:** Non-atomic write operations in RibbitCache::write() - metadata and
data files written separately, causing race conditions during concurrent access

**Solution:** Implement atomic writes using temporary files:
- Write to .tmp files in same directory as final files
- Use rename() for atomic promotion to final filenames
- Add proper error cleanup for temporary files
- Ensure both data and metadata appear simultaneously

**Result:** Test now passes consistently in all environments

### Quality Improvements

**Clippy Fixes:**
- Update format strings to use inline variable syntax
- Replace multiple .replace() calls with single array-based replacement
- Use is_some_and() instead of map_or(false, ...) for better idioms

**All QA Checks Pass:**
- ✅ cargo fmt --all -- --check
- ✅ cargo clippy --all-features --all-targets -- -D warnings
- ✅ cargo test --all-features --workspace (577+ tests passing)
- ✅ RUSTDOCFLAGS="-D warnings" cargo doc --all-features --no-deps
- ✅ cargo build --release --workspace

### Testing

- 100% test pass rate including previously flaky cache test
- New comprehensive test suite for TACT manifest integration
- All integration tests passing with real WoW data formats
- Race condition eliminated with consistent atomic operations

+++ casc-storage/src/index/group_index.rs
@@ -48,17 +48,17 @@ impl GroupIndex {
         );
 
         // Calculate entry size
-        let entry_size = header.ekey_bytes
-            + header.archive_bytes
-            + header.span_offset_bytes
-            + header.span_size_bytes;
+        let entry_size = (header.ekey_bytes as usize)
+            + (header.archive_bytes as usize)
+            + (header.span_offset_bytes as usize)
+            + (header.span_size_bytes as usize);
 
         // Read all remaining data
         let mut data = Vec::new();
         reader.read_to_end(&mut data)?;
 
         // Parse entries
-        let entry_count = data.len() / entry_size as usize;
+        let entry_count = data.len() / entry_size;
         debug!(
             "Parsing {} entries of {} bytes each",
             entry_count, entry_size
@@ -69,7 +69,7 @@ impl GroupIndex {
 
         for i in 0..entry_count {
             let entry = Self::parse_entry(&data[offset..], &header)?;
-            offset += entry_size as usize;
+            offset += entry_size;
 
             if i < 5 {
                 trace!(
